mysite/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
django-debug-toolbar
psycopg2-binary